requirements.txt
requests
simplejson
psycopg2-binary
